+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Not released yet
 
+## 0.9.0 (2023-10-09)
+
 * Add `fingerprint()` function to condition code execution based on some hash changes
 * Better handle default Symfony commands when no castor file exists yet
 * Add `-c` option to `castor` command to specify a context